Add files via upload
Uploading in groups of the scenes, assets, and scripts. I admit I have never used github before, if I am doing this wrong please let me know and I can correct it asap.

Uploaded should be a main menu scene, a pause menu asset that can go in the active game scene, and a game over scene that allows the player to restart the game by going back to the main menu.